+++ tests/integration/conftest.py
@@ -13,6 +13,7 @@ import pytest
 import pytest_asyncio
 import yaml
 from github import Github, GithubException
+from github.Branch import Branch
 from github.Repository import Repository
 from juju.application import Application
 from juju.model import Model
@@ -293,3 +294,37 @@ async def app_with_forked_repo(
     await ensure_charm_has_runner(app=app, model=model)
 
     yield app
+
+
+@pytest.fixture(scope="module")
+def forked_github_branch(
+    github_repository: Repository, forked_github_repository: Repository
+) -> Iterator[Branch]:
+    """Create a new forked branch for testing.
+
+    This branch should only be used for this module.
+    """
+    branch_name = f"test/{secrets.token_hex(4)}"
+
+    main_branch = forked_github_repository.get_branch(github_repository.default_branch)
+    branch_ref = forked_github_repository.create_git_ref(
+        ref=f"refs/heads/{branch_name}", sha=main_branch.commit.sha
+    )
+
+    for _ in range(10):
+        try:
+            branch = forked_github_repository.get_branch(branch_name)
+            break
+        except GithubException as err:
+            if err.status == 404:
+                sleep(5)
+                continue
+            raise
+    else:
+        assert (
+            False
+        ), "Failed to get created branch in fork repo, the issue with GitHub or network."
+
+    yield branch
+
+    branch_ref.delete()
